split deposit
