+++ src/components/wallet/Wallet.tsx
@@ -1,48 +1,20 @@
 import * as React from 'react'
 import { connect } from 'react-redux'
-import { Dispatch } from 'redux'
-import { ThunkDispatch } from 'redux-thunk'
 import { AppState } from '../../redux/modules'
-import {
-  DEPOSIT_STATUS,
-  State as DepositState,
-  deposit
-} from '../../redux/modules/chamberWallet/deposit'
 import {
   WALLET_STATUS,
   State as WalletState
 } from '../../redux/modules/chamberWallet/wallet'
 import UTXOList from './UTXOList'
-import { Button, LoadingSpinner } from '../common'
-import { MarginHorizontal } from '../utility'
-import {
-  FONT_SIZE,
-  PADDING,
-  BORDER,
-  MARGIN,
-  BOX_SHADOW
-} from '../../constants/size'
+import { FONT_SIZE, PADDING, MARGIN, BOX_SHADOW } from '../../constants/size'
 import colors from '../../constants/colors'
 import Link from 'next/link'
 
 interface StateProps {
   wallet: WalletState
-  depositState: DepositState
 }
 
-interface DispatchProps {
-  deposit: (ether: number) => void
-}
-
-interface State {
-  depositAmount: number
-}
-
-class WalletCard extends React.Component<StateProps & DispatchProps, State> {
-  public state = {
-    depositAmount: 1
-  }
-
+class WalletCard extends React.Component<StateProps> {
   public async componentDidMount() {
     const { ref } = this.props.wallet
 
@@ -68,7 +40,6 @@ class WalletCard extends React.Component<StateProps & DispatchProps, State> {
 
     const { ref } = wallet
     const balance = ref.getBalance()
-    const depositStatus = this.props.depositState.status
     const utxos = ref.getUTXOArray()
 
     return (
@@ -93,57 +64,25 @@ class WalletCard extends React.Component<StateProps & DispatchProps, State> {
               </Link>
             </div>
             <div className="link-card">
-              <Link prefetch href="/" /* TODO: implement */>
+              <Link prefetch href="/deposit">
                 <button className="link-button">Deposit</button>
               </Link>
             </div>
           </section>
           {/* UTXOList section */}
           <UTXOList utxos={utxos} wallet={ref} />
-          <section className="control-section">
-            {/* Control section */}
-            <h3 className="control-title">DEPOSIT</h3>
-            <div className="deposit-control">
-              <div>
-                <select
-                  className="deposit-amount-select"
-                  onChange={this.handleChangeDepositAmount}
-                >
-                  <option value={1}>1 ETH</option>
-                  <option value={2}>2 ETH</option>
-                  <option value={5}>5 ETH</option>
-                  <option value={10}>10 ETH</option>
-                </select>
-              </div>
-              <div className="controls">
-                <Button
-                  disabled={depositStatus === DEPOSIT_STATUS.LOADING}
-                  onClick={this.handleDeposit}
-                >
-                  Deposit
-                </Button>
-                {depositStatus === DEPOSIT_STATUS.LOADING && (
-                  <>
-                    <MarginHorizontal />
-                    <LoadingSpinner size="medium" />
-                  </>
-                )}
-              </div>
-            </div>
-          </section>
         </div>
-        {/* TODO: place Transfer link */}
         <style jsx>{`
           .body {
             margin: ${MARGIN.LARGE};
-            color: ${colors.TEXT_MAIN};
           }
 
           .heading {
             background-color: ${colors.BG_WHITE};
-            color: ${colors.TEXT_MAIN};
             box-shadow: ${BOX_SHADOW.VERTICAL_NORMAL};
             padding: 0 ${PADDING.VERY_LARGE} ${PADDING.LARGE};
+            height: 24rem;
+            transition: height 0.4s linear;
           }
 
           .address-section {
@@ -194,58 +133,14 @@ class WalletCard extends React.Component<StateProps & DispatchProps, State> {
             font-size: ${FONT_SIZE.MEDIUM};
             text-transform: uppercase;
           }
-
-          .control-section {
-            padding-top: ${PADDING.MEDIUM};
-            border-top: solid ${BORDER.THICK} ${colors.BORDER_COLOR_LIGHT};
-          }
-
-          .deposit-control {
-            display: flex;
-            align-items: center;
-          }
-
-          .deposit-control > .controls {
-            display: flex;
-            align-items: center;
-          }
-
-          .deposit-amount-select {
-            color: ${colors.TEXT_MAIN};
-            background-color: transparent;
-            border: solid ${BORDER.THICK} ${colors.BORDER_COLOR_LIGHT};
-            font-size: ${FONT_SIZE.MEDIUM};
-            margin-right: ${MARGIN.MEDIUM};
-            padding: ${PADDING.MEDIUM};
-          }
-
-          .deposit-title {
-            font-size: ${FONT_SIZE.MEDIUM};
-          }
         `}</style>
       </main>
     )
-  }
-
-  private handleDeposit = () => {
-    this.props.deposit(this.state.depositAmount)
-  }
-
-  private handleChangeDepositAmount = (
-    e: React.ChangeEvent<HTMLSelectElement>
-  ) => {
-    this.setState({ depositAmount: Number(e.target.value) })
   }
 }
 
 export default connect(
   (state: AppState): StateProps => ({
-    wallet: state.chamberWallet.wallet,
-    depositState: state.chamberWallet.deposit
-  }),
-  (dispatch: Dispatch): DispatchProps => ({
-    deposit: (ether: number) => {
-      ;(dispatch as ThunkDispatch<void, AppState, any>)(deposit(ether))
-    }
+    wallet: state.chamberWallet.wallet
   })
 )(WalletCard)
